refactor: migration sources file & add type source
