+++ news-api/src/index.html
@@ -1,4 +1,4 @@
-<!DOCTYPE html>
+<!doctype html>
 <html>
     <head>
         <meta charset="utf-8" />
@@ -15,9 +15,7 @@
             <div class="news"></div>
         </main>
         <footer>
-            <p class="copyright">
-                Copyright <a href="https://newsapi.org">NewsAPI</a>
-            </p>
+            <p class="copyright">Copyright <a href="https://newsapi.org">NewsAPI</a></p>
         </footer>
 
         <template id="sourceItemTemp">
